Service implemented for Actors, crud operations functional for Actors and Films. working now on service for Films

+++ src/main/java/com/example/sakila/dto/request/FilmRequest.java
@@ -14,8 +14,8 @@ public class FilmRequest {
     private final String title;
     private final String description;
     private final Year release_year;
-    private final Short language_id;
-    private final Short rental_duration;
+    private final Byte language_id;
+    private final Byte rental_duration;
     private final Float rental_rate;
     private final Short length;
     private final Float replacement_cost;
